Combined the local routing convention into LocalTransport for easier changes later
